User 실패 테스트 코드 작성 (#26)
* test: 코드 위치 수정

* test: 수정 중

* test: 유저 추가 실패 테스트
